re-organize seed scripts
