+++ src/lib/seed/data/assets.ts
